+++ server/src/modules/factions/factions.service.ts
@@ -187,8 +187,7 @@ export class FactionsService {
       }
     } else {
       // No user progress provided, get default or most recent
-      selectedMedia =
-        allMedia.data.find((media) => media.isDefault) || allMedia.data[0];
+      selectedMedia = allMedia.data[0];
     }
 
     if (selectedMedia) {