+++ docs/CHANGELOG.md
@@ -3,13 +3,15 @@
 - Redis based rate limiting with caching middle ware for responses; add cache headers to specific routes
 - Update factions page/data; include members and possibly calculate faction switches based on events
 - Add highlight of most popular quote, gamble, and character profile picture
-- Add different image profile pictures for characters and adding labeling (update character image based on progressed chapter; subtle chapter progression)
+- Allow profiles to set their profile image based on entity display media for any character
 - Add image upload for gambles, events, factions, volumes (need to setup Backblaze + CDN); allow character profile picture choice from characters
 - Update landing page volume showcase to handle other volume covers (single and pairs) 
 - Add proper README at the root of the project; add AGPL v3 license to allow people to copy but force them to open source as well
 
 ## 2025-09-11
 - Overhauled media to now support polymorphic ownership; added entity purpose
+- Added media thumbnails to load images
+- Implemented media with entity display type to load based on chapter progress and update the thumbnail
 
 ## 2025-09-10
 ### Changes
@@ -17,8 +19,6 @@
 - Fixed guides page to show descriptions and link user profiles
 - Added confirmation for deletion of data
 
-### Notes
-- 
 
 ## 2025-09-09
 ### Changes